Silence most warnings
Suppress and fix many clang, GCC and MSVC warnings

"cause the night is crap and full of warnings"

+++ chromium/base/allocator/partition_allocator/src/partition_alloc/page_allocator_constants.h
@@ -159,7 +159,7 @@ SystemPageSize() {
   // on a non-constexpr.
   return PageAllocationGranularity();
 #else
-  return 1 << SystemPageShift();
+  return size_t(1) << SystemPageShift();
 #endif
 }
